% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{largedemodata}
\alias{largedemodata}
\title{Large demo dataset}
\format{A data frame with 82,799 rows and 2 variables:
\describe{
  \item{x}{seconds}
  \item{obs}{observed measurement}
}}
\source{
\url{https://sites.google.com/site/dnartrahckcir/home/tvdiff-code}
}
\usage{
largedemodata
}
\description{
"Data obtained from a whole-room calorimeter, courtesy of Edward L. Melanson
of the University of Colorado Denver. The metabolic rate of a subject within the
calorimeter can be determined via respirometry, the measurement of oxygen consumption,
and carbon dioxide production within the room." "The data consists of samples
taken every second for most of a day, for a total of 82,799 samples."
}
\references{
Rick Chartrand, "Numerical differentiation of noisy, nonsmooth
data," ISRN Applied Mathematics, Vol. 2011, Article ID 164564, 2011
}
\keyword{datasets}
