% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{paleo}
\alias{paleo}
\title{Paleo diatom abundances}
\format{A data frame with 83,167 rows and 5 variables:
\describe{
  \item{site}{seconds}
  \item{sortVar}{time}
  \item{variable}{species}
  \item{value}{relative abundance}
  \item{cellID}{id variable}
}}
\source{
\url{https://doi.org/10.1371/journal.pone.0108936.s001}
}
\usage{
paleo
}
\description{
"To test for regime shifts in the paleoecological record, we used a long-term
high-resolution sedimentological record from Foy Lake (Montana, USA) that
showed abrupt changes in diatom community structure at 1.3 ka (thousands
of years before present, with present defined as AD 1950). Foy Lake
(48.1648N, 1143589W, 1005 m elevation) is a deep freshwater lake situated
in the drought-sensitive Flathead River Basin in the Northern Rocky Mountains
[4], [5]. Diatom assemblages in this system are sensitive to changes in
lake depth driven by changes in effective moisture [6] and represent
one metric of ecological resilience. The percent abundances of 109 diatom
species were collected from a lake sediment core that was sampled
continuously at an interval of every 5-20 years, yielding a 7 kyr
record of 800 time-steps."
}
\references{
Spanbauer, T. L., Allen, C. R., Angeler, D. G., Eason, T., Fritz,
S. C., Garmestani, A. S., … Stone, J. R. (2014). Prolonged Instability Prior
to a Regime Shift. PLOS ONE, 9(10), e108936.
https://doi.org/10.1371/journal.pone.0108936
}
\keyword{datasets}
