% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pcg.R
\name{pcg}
\alias{pcg}
\title{Preconditioned conjugate gradient method solver}
\usage{
pcg(Ax, b, M, x0, maxiter = 1000, tol = 1e-06)
}
\arguments{
\item{Ax}{function that takes argument x and returns matrix product A*x}

\item{b}{right hand side of the linear system}

\item{M}{preconditioner for A}

\item{x0}{starting guess for solution}

\item{maxiter}{maximum number of iterations}

\item{tol}{tolerance for convergence on norm(residual, "2")}
}
\description{
Solve system of linear equations $Ax = b$
}
\examples{
A <- matrix(c(4, 1, 1, 3), nrow = 2)
b <- c(1, 2)
x0 <- c(2, 1)

Ax <- function(x) {
  A \%*\% x
}

M <- matrix(c(4, 0, 0, 3), nrow = nrow(A))

opt <- pcg(Ax, b, M, x0)

}
