% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pcgsolve}
\alias{pcgsolve}
\title{Preconditioned conjugate gradient method}
\usage{
pcgsolve(b, M, alph, L, dx, tol = 1e-06, maxIter = 1000L)
}
\arguments{
\item{b}{vector, with same dimension as number of rows of A.}

\item{M}{matrix, preconditioner matrix defined interal to \code{\link{TVRegDiffR}}.}

\item{alph}{numeric, regularization parameter used in \code{\link{TVRegDiffR}}.}

\item{L}{matrix, linearized diffusion matrix internal to \code{\link{TVRegDiffR}}.}

\item{dx}{numeric, grid spacing used in \code{\link{TVRegDiffR}}.}

\item{tol}{numeric, threshold for convergence, default is \code{1e-6}.}

\item{maxIter}{numeric, maximum iteration, default is \code{1000}.}
}
\value{
A vector representing solution x.
}
\description{
Preconditioned conjugate gradient method for solving system of linear equations Ax = b,
where A is symmetric and positive definite.
}
\details{
Code is a slightly modified version of \code{\link[cPCG]{pcgsolve}}
}
