% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TVRegDiffR.R
\name{TVRegDiffR}
\alias{TVRegDiffR}
\title{Total Variation Regularized Numerical Differentiation (TVDiff)}
\usage{
TVRegDiffR(data, iter, alph, u0 = NULL, scale = "small", ep = 1e-06,
  dx, plotflag = 0, diagflag = 0, tol = 1e-04, maxit = 200)
}
\arguments{
\item{data}{Vector of data to be differentiated.}

\item{iter}{Number of iterations to run the main loop. A stopping condition
based on the norm of the gradient vector g below would be an easy
modification.  No default value.}

\item{alph}{Regularization parameter.  This is the main parameter
to fiddle with.  Start by varying by orders of magnitude until reasonable
results are obtained.  A value to the nearest power of 10 is usally adequate.
No default value.  Higher values increase regularization strenght and improve
conditioning.}

\item{u0}{Initialization of the iteration.  Default value is the naive
derivative (without scaling), of appropriate length (this being different
for the two methods). Although the solution is theoretically independent of
the intialization, a poor choice can exacerbate conditioning issues when the
linear system is solved.}

\item{scale}{'large' or 'small' (case insensitive).  Default is 'small'.
'small' has somewhat better boundary behavior, but becomes unwieldly for data
larger than 1000 entries or so.  'large' has simpler numerics but is more
efficient for large-scale problems.  'large' is more readily modified for
higher-order derivatives, since the implicit differentiation matrix is square.}

\item{ep}{Parameter for avoiding division by zero.  Default value is 1e-6.
Results should not be very sensitive to the value.  Larger values improve
conditioning and therefore speed, while smaller values give more accurate
results with sharper jumps.}

\item{dx}{Grid spacing, used in the definition of the derivative operators.
Default is the reciprocal of the data size.}

\item{plotflag}{Flag whether to display plot at each iteration. Default is 1
(yes).  Useful, but adds significant running time.}

\item{diagflag}{Flag whether to display diagnostics at each iteration.
Default is 1 (yes).  Useful for diagnosing preconditioning problems.  When
tolerance is not met, an early iterate being best is more worrying than a
large relative residual.}

\item{tol}{R Version Only: Tolerance passed to scipy.sparse.linalg.cg}

\item{maxit}{R Version Only: Maximum iterations passed to
scipy.sparse.linalg.cg}
}
\value{
Estimate of the regularized derivative of data.  Due to different
grid assumptions, length( u ) = length( data ) + 1 if scale = 'small',
otherwise length( u ) = length( data ).
}
\description{
Estimate the derivative of noisy data using regularized differentiation.
}
\details{
This function is an R wrapper for a Python translation of the original Matlab
code.
}
\references{
Rick Chartrand, "Numerical differentiation of noisy, nonsmooth
data," ISRN Applied Mathematics, Vol. 2011, Article ID 164564, 2011.
}
\author{
Rick Chartrand (\email{rickc@lanl.gov})

R translation: Nathaniel Price  (\email{natbprice@gmail.com})
}
